Crate azuki_tac

TAC (Three-Address Code) is the intermediate representation for Azuki programs. This code type is similar to Cranelift IR and LLVM IR.

This crate contains definition of Azuki TAC, code to serialize and deserialize them, and to construct them from regular control flows in a higher-level programming language.

See the details in separate modules.

Re-exports

pub use ty::NumericTy;

pub use ty::Ty;

pub use ty::TyKind;

pub use containers::BBId;

pub use containers::InstId;

Modules

builder
A builder for constructing SSA functions from regular control flows.

containers

err
Error and Result types.

formatter
Serialization and de-serialization for TAC code.

linkedlist
Some random linked list traits, may or may not be used in actual program.

optimizer

parser

ty
Type system definitions and stuff.

util

Structs

BasicBlock
A single basic block, represented as an indirect doubly linked list of instructions.

BinaryInst

FunctionCall

Inst

ItemsIter
An iterator of all items in one chain in an ImplicitLinkedList. Created by [ImplicitLinkedList::item_iter].

Program

Tac
Represents a single TAC instruction inside an indirect doubly linked list of instructions.

TacFunc
A function made of TAC instructions.

Enums

BinaryOp

Branch
Represents a branch instruction.

InstKind
Kinds of an instruction

Value

Traits

ImplicitLinkedList
An implicitly linked list, allowing freestanding items and multiple links inside itself.

ImplicitLinkedListGet2

ImplicitLinkedListItem
An implicit linked list item. Contains keys for the previous and next items of Self.

Type Definitions

Immediate