Struct azuki_tac::builder::FuncBuilder

pub struct FuncBuilder<'a, TVar> {
    pub editor: FuncEditor<'a>,
    sealed_bbs: BitSet,
    filled_bbs: BitSet,
    variable_map: BTreeMap<TVar, (Ty, BTreeMap<BBId, Index>)>,
    incomplete_phi: BTreeMap<BBId, Vec<(TVar, Index)>>,
    pub block_pred: BTreeMap<BBId, TinyVec<[BBId; 7]>>,
}

A function builder that loosely resembles building SSA functions using the algorithm described in Simple and Efficient Construction of Static Single Assignment Form by Matthias Braun et al.

This type is parameterized by one type, TVar, which is the representation of variable in your original language. It is expected to be a small, Clone Ord and Debug type for usage inside BTreeMaps

Fields
editor: FuncEditor<'a>

The function we're building.

sealed_bbs: BitSet

Sealed basic blocks.

Sealed basic blocks have all their predecessors determined.

filled_bbs: BitSet

Filled basic blocks.

Filled basic blocks have finished filling in calculation instructions.

variable_map: BTreeMap<TVar, (Ty, BTreeMap<BBId, Index>)>

A Variable - Basic Block - SSA Value map.

This map is for global and local value numbering in the algorithm. Since we expect variables and basic blocks to be small integer IDs, BTreeMap is preferred here over HashMap.

Well... You see, there's a forest in cranelift, right?

incomplete_phi: BTreeMap<BBId, Vec<(TVar, Index)>>

Incomplete phi commands (params in our case).

block_pred: BTreeMap<BBId, TinyVec<[BBId; 7]>>

Basic blocks predecessors

Implementations

impl<'a, TVar> FuncBuilder<'a, TVar> where
TVar: Ord + Debug + Clone,

pub fn new(func_editor: FuncEditor<'a>) -> FuncBuilder<'a, TVar>

Create a function builder for a function with the given name and type undefined for now.

pub fn new_func(func: &'a mut TacFunc) -> FuncBuilder<'a, TVar>

Create a function builder for a function with the given name and given type ty.

pub fn sanity_check(self)

Build this function.

Panics

This function panics when there is any basic block not filled, not sealed, or there is any incomplete phis lying around.

pub fn mark_sealed(&mut self, bb_id: BBId)

Mark the given basic block as sealed. Also completes all incomplete Phi commands inside this basic block.

Sealed blocks have all its predecessors determined.

pub fn mark_filled(&mut self, bb_id: BBId)

Mark the given basic block as filled.

Filled blocks have all its instructions inserted.

pub fn is_sealed(&self, bb_id: BBId) -> bool

Check if the given basic block is sealed.

pub fn is_filled(&self, bb_id: BBId) -> bool

Check if the given basic block is filled.

pub fn declare_var(&mut self, var: TVar, ty: Ty)

pub fn write_variable_cur(&mut self, var: TVar, inst: Index) -> TacResult<()>

Indicate that variable var is written as the result of instruction inst.

pub fn write_variable(
&mut self,
var: TVar,
inst: Index,
bb_id: BBId
) -> TacResult<()>

Indicate that variable var is written as the result of instruction inst in basic block bb_id. If the variable does not exist, it will be created.

pub fn read_variable_cur(&mut self, var: TVar) -> Option<Index>

Indicate that variable var is read in the current basic block. Returns the index to the latest definition of this variable, or None if it does not exist.

Side effects

According to the algorithm, this function may introduce parameters to bb and insert parameter passes to the block's predecessors.

pub fn read_variable(&mut self, var: TVar, bb_id: BBId) -> Option<Index>

Indicate that variable var is read in basic block bb. Returns the index to the latest definition of this variable, or None if it does not exist.

Side effects

According to the algorithm, this function may introduce parameters to bb and insert parameter passes to the block's predecessors.

fn read_variable_recursive(&mut self, var: TVar, bb_id: BBId) -> Option<Index>

This function directly corresponds to readVariableRecursive in the algorithm.

fn add_phi_operands(&mut self, var: TVar, phi: Index, preds: &[BBId])

This function directly corresponds to addPhiOperands in the algorithm.

fn try_remove_trivial_phi(&mut self, phi_op: Index)

pub fn add_branch(&mut self, from: BBId, to: BBId)

pub fn pred_of_bb(&self, bb_id: BBId) -> TinyVec<[BBId; 7]>

Returns an iterator of all predecessors of a basic block.

The return type is to make the borrow checker happy.

pub fn succ_of_bb(&self, bb_id: BBId) -> impl Iterator<Item = BBId> + '_

Returns an iterator of all successors of a basic block.