Enum azuki_tac::InstKind

pub enum InstKind {
    Binary(BinaryInst),
    FunctionCall(FunctionCall),
    Assign(Value),
    Phi(BTreeMap<BBId, InstId>),
    Param(usize),
    Dead,
}

Kinds of an instruction

Variants

Binary(BinaryInst)

A binary operaton, e.g. plus, divide

FunctionCall(FunctionCall)

A call to another function.

Assign(Value)

An assignment from another instruction or constant

Phi(BTreeMap<BBId, InstId>)

A phi instruction

Param(usize)

A function parameter

Dead

An unreachable value

Implementations

impl InstKind

pub fn as_binary_mut(&mut self) -> Option<&mut BinaryInst>

Optionally returns mutable references to the inner fields if this is a InstKind::Binary, otherwise None

pub fn as_binary(&self) -> Option<&BinaryInst>

Optionally returns references to the inner fields if this is a InstKind::Binary, otherwise None

pub fn into_binary(self) -> Result<BinaryInst, Self>

Returns the inner fields if this is a InstKind::Binary, otherwise returns back the enum in the Err case of the result

pub fn as_function_call_mut(&mut self) -> Option<&mut FunctionCall>

Optionally returns mutable references to the inner fields if this is a InstKind::FunctionCall, otherwise None

pub fn as_function_call(&self) -> Option<&FunctionCall>

Optionally returns references to the inner fields if this is a InstKind::FunctionCall, otherwise None

pub fn into_function_call(self) -> Result<FunctionCall, Self>

Returns the inner fields if this is a InstKind::FunctionCall, otherwise returns back the enum in the Err case of the result

pub fn as_assign_mut(&mut self) -> Option<&mut Value>

Optionally returns mutable references to the inner fields if this is a InstKind::Assign, otherwise None

pub fn as_assign(&self) -> Option<&Value>

Optionally returns references to the inner fields if this is a InstKind::Assign, otherwise None

pub fn into_assign(self) -> Result<Value, Self>

Returns the inner fields if this is a InstKind::Assign, otherwise returns back the enum in the Err case of the result

pub fn as_phi_mut(&mut self) -> Option<&mut BTreeMap<BBId, InstId>>

Optionally returns mutable references to the inner fields if this is a InstKind::Phi, otherwise None

pub fn as_phi(&self) -> Option<&BTreeMap<BBId, InstId>>

Optionally returns references to the inner fields if this is a InstKind::Phi, otherwise None

pub fn into_phi(self) -> Result<BTreeMap<BBId, InstId>, Self>

Returns the inner fields if this is a InstKind::Phi, otherwise returns back the enum in the Err case of the result

pub fn as_param_mut(&mut self) -> Option<&mut usize>

Optionally returns mutable references to the inner fields if this is a InstKind::Param, otherwise None

pub fn as_param(&self) -> Option<&usize>

Optionally returns references to the inner fields if this is a InstKind::Param, otherwise None

pub fn into_param(self) -> Result<usize, Self>

Returns the inner fields if this is a InstKind::Param, otherwise returns back the enum in the Err case of the result

pub fn as_dead_mut(&mut self) -> Option<()>

Optionally returns references to the inner fields if this is a InstKind::Dead, otherwise None

pub fn as_dead(&self) -> Option<()>

Optionally returns references to the inner fields if this is a InstKind::Dead, otherwise None

impl InstKind

pub fn params_iter(&self) -> impl Iterator<Item = Value> + '_

pub fn param_op_iter(&self) -> impl Iterator<Item = InstId> + '_