Trait azuki_tac::linkedlist::ImplicitLinkedList

pub trait ImplicitLinkedList<Key> where
    Key: Copy + Eq, {
    type Item: ImplicitLinkedListItem<Key = Key>;
    fn get_item(&self, key: Key) -> &Self::Item;

    fn get_item_mut(&mut self, key: Key) -> &mut Self::Item;

    fn insert_item(&mut self, item: Self::Item) -> Key;

    fn remove_item(&mut self, idx: Key) -> Self::Item;

    fn next_key(&self, key: Key) -> Option<Key> { ... }

    fn prev_key(&self, key: Key) -> Option<Key> { ... }

    fn attach_after(&mut self, after: Key, this: Key) { ... }

    fn attach_before(&mut self, before: Key, this: Key) { ... }

    fn detach(&mut self, idx: Key) { ... }

    fn split_after(&mut self, idx: Key) -> Option<Key> { ... }

    fn split_before(&mut self, idx: Key) -> Option<Key> { ... }

    fn connect(&mut self, tail: Key, head: Key) { ... }

    fn items_iter(
        &self, 
        from: Option<Key>, 
        to: Option<Key>
    ) -> ItemsIter<'_, Self, Key> { ... }
}

An implicitly linked list, allowing freestanding items and multiple links inside itself.

Associated Types

type Item: ImplicitLinkedListItem<Key = Key>

Loading content...
Required methods

fn get_item(&self, key: Key) -> &Self::Item

fn get_item_mut(&mut self, key: Key) -> &mut Self::Item

fn insert_item(&mut self, item: Self::Item) -> Key

fn remove_item(&mut self, idx: Key) -> Self::Item

Loading content...
Provided methods

fn next_key(&self, key: Key) -> Option<Key>

fn prev_key(&self, key: Key) -> Option<Key>

fn attach_after(&mut self, after: Key, this: Key)

Position this item after the given item.

fn attach_before(&mut self, before: Key, this: Key)

Position this item before the given item.

fn detach(&mut self, idx: Key)

Detaches this item from the list.

fn split_after(&mut self, idx: Key) -> Option<Key>

Split the chain into two after the given item, or return None if no item is after the given item.

fn split_before(&mut self, idx: Key) -> Option<Key>

Split the chain into two before the given item, or return None if no item is before the given item.

fn connect(&mut self, tail: Key, head: Key)

fn items_iter(
&self,
from: Option<Key>,
to: Option<Key>
) -> ItemsIter<'_, Self, Key>

Creates an iterator of items of Self in a chain, from item from to to.

If from is None, the resulting iterator will be empty.

If to is None, the resulting iterator will iterate until the end of chain.

Loading content...