Trait azuki_tac::optimizer::FunctionOptimizer

pub trait FunctionOptimizer {
    fn name(&self) -> Cow<'_, str>;

    fn edits_program(&self) -> bool;

    fn optimize_func(
        &mut self, 
        env: &mut OptimizeEnvironment, 
        func: &mut TacFunc
    );

    fn reset(&mut self) { ... }

    fn do_initialization(
        &mut self, 
        _env: &mut OptimizeEnvironment, 
        _prog: &Program
    ) { ... }

    fn do_finalization(
        &mut self, 
        _env: &mut OptimizeEnvironment, 
        _prog: &Program
    ) { ... }

    fn make_pass(self) -> FunctionOptimizerPass<Self>
    where
        Self: Sized,
    { ... }
}

Required methods

fn name(&self) -> Cow<'_, str>

Returns the name of this pass.

fn edits_program(&self) -> bool

Whether this pass edits the program. This method should correctly represent the function of [optimize_program], since mutable references are given anyway.

fn optimize_func(&mut self, env: &mut OptimizeEnvironment, func: &mut TacFunc)

Optimize a single function.

Loading content...
Provided methods

fn reset(&mut self)

Reset this instance for optimizing another function.

fn do_initialization(&mut self, _env: &mut OptimizeEnvironment, _prog: &Program)

Perform initialization before any functions are processed.

fn do_finalization(&mut self, _env: &mut OptimizeEnvironment, _prog: &Program)

Perform finalization after all functions are processed.

fn make_pass(self) -> FunctionOptimizerPass<Self> where
Self: Sized,

Transform this optimizer into a pass. You should not overwrite this method in most cases.

Loading content...