Trait azuki_tac::optimizer::Pass

pub trait Pass {
    fn name(&self) -> Cow<'_, str>;

    fn edits_program(&self) -> bool;

    fn optimize_program(
        &mut self, 
        env: &mut OptimizeEnvironment, 
        program: &mut Program
    );
}

Represents a single pass inside the compilation pipeline.

A Pass should be constructible from an OptimizeEnvironment, which supplies external data needed for this pass.

Required methods

fn name(&self) -> Cow<'_, str>

Returns the name of this pass.

fn edits_program(&self) -> bool

Whether this pass edits the program. This method should correctly represent the function of [optimize_program], since mutable references are given anyway.

fn optimize_program(
&mut self,
env: &mut OptimizeEnvironment,
program: &mut Program
)

Optimize at program level.

Loading content...