Module azuki_tac::parser

Modules

easy_parse
Stream wrapper which provides an informative and easy to use error type.

implementation
Module for parsing the text representation of Azuki TAC into real code.

parse_stream
Streams are similar to the Iterator trait in that they represent some sequential set of items which can be retrieved one by one. Where Streams differ is that they are allowed to return errors instead of just None and if they implement the RangeStreamOnce trait they are also capable of returning multiple items at the same time, usually in the form of a slice.

Traits

EasyParser
Provides the easy_parse method which provides good error messages by default

Parser
By implementing the Parser trait a type says that it can be used to parse an input stream into the type Output.

Stream
A stream of tokens which can be duplicated

Functions

parse_func

parse_program

parse_program_from_string