Struct azuki_tac::parser::implementation::parse_stream::buffered::Stream

pub struct Stream<Input> where
    Input: StreamOnce + Positioned,  {
    offset: usize,
    iter: Input,
    buffer_offset: usize,
    buffer: VecDeque<(<Input as StreamOnce>::Token, <Input as StreamOnce>::Position)>,
}

Stream which buffers items from an instance of StreamOnce into a ring buffer. Instances of StreamOnce which is not able to implement ResetStream (such as ReadStream) may use this as a way to implement ResetStream and become a full Stream instance.

The drawback is that the buffer only stores a limited number of items which limits how many tokens that can be reset and replayed. If a buffered::Stream is reset past this limit an error will be returned when uncons is next called.

NOTE: If this stream is used in conjunction with an error enhancing stream such as easy::Stream (also via the easy_parser method) it is recommended that the buffered::Stream instance wraps the easy::Stream instance instead of the other way around.

// DO
buffered::Stream::new(easy::Stream(..), ..)
// DON'T
easy::Stream(buffered::Stream::new(.., ..))
parser.easy_parse(buffered::Stream::new(..));

Fields
offset: usizeiter: Inputbuffer_offset: usizebuffer: VecDeque<(<Input as StreamOnce>::Token, <Input as StreamOnce>::Position)>
Implementations

impl<Input> Stream<Input> where
Input: StreamOnce + Positioned,
<Input as StreamOnce>::Position: Clone,
<Input as StreamOnce>::Token: Clone,

pub fn new(iter: Input, lookahead: usize) -> Stream<Input>

Constructs a new BufferedStream from a StreamOnce instance with a lookahead number of elements that can be stored in the buffer.