Trait azuki_tac::parser::parse_stream::ResetStream

pub trait ResetStream: StreamOnce {
    type Checkpoint: Clone;
    pub fn checkpoint(&self) -> Self::Checkpoint;

    pub fn reset(
        &mut self, 
        checkpoint: Self::Checkpoint
    ) -> Result<(), Self::Error>;
}

A StreamOnce which can create checkpoints which the stream can be reset to

Associated Types

type Checkpoint: Clone

Loading content...
Required methods

pub fn checkpoint(&self) -> Self::Checkpoint

Creates a Checkpoint at the current position which can be used to reset the stream later to the current position

pub fn reset(&mut self, checkpoint: Self::Checkpoint) -> Result<(), Self::Error>

Attempts to reset the stream to an earlier position.

Loading content...