Trait azuki_tac::parser::parse_stream::StreamOnce

pub trait StreamOnce {
    type Token: Clone;
    type Range: Clone;
    type Position: Clone + Ord;
    type Error: ParseError<Self::Token, Self::Range, Self::Position>;
    pub fn uncons(
        &mut self
    ) -> Result<Self::Token, <Self::Error as ParseError<Self::Token, Self::Range, Self::Position>>::StreamError>;

    pub fn is_partial(&self) -> bool { ... }
}

StreamOnce represents a sequence of items that can be extracted one by one.

Associated Types

type Token: Clone

The type of items which is yielded from this stream.

type Range: Clone

The type of a range of items yielded from this stream. Types which do not a have a way of yielding ranges of items should just use the Self::Token for this type.

type Position: Clone + Ord

Type which represents the position in a stream. Ord is required to allow parsers to determine which of two positions are further ahead.

type Error: ParseError<Self::Token, Self::Range, Self::Position>

Loading content...
Required methods

pub fn uncons(
&mut self
) -> Result<Self::Token, <Self::Error as ParseError<Self::Token, Self::Range, Self::Position>>::StreamError>

Takes a stream and removes its first token, yielding the token and the rest of the elements. Returns Err if no element could be retrieved.

Loading content...
Provided methods

pub fn is_partial(&self) -> bool

Returns true if this stream only contains partial input.

See PartialStream.

Loading content...