Struct azuki_tac::TacFunc

pub struct TacFunc {
    pub name: SmolStr,
    pub ty: Ty,
    instructions_arena: SlotMap<InstId, Tac>,
    basic_block_arena: SlotMap<BBId, BasicBlock>,
    pub first_block: Option<BBId>,
}

A function made of TAC instructions.

The instructions are represented as an indirect doubly linked list inside the arena using item indices. Every basic block holds the start and end index of its instructions.

ALL methods of this type will panic if you feed invalid indices into them. If that happens, there's definitely bug inside your code.

Fields
name: SmolStr

Function name

ty: Ty

Function type

instructions_arena: SlotMap<InstId, Tac>

An arena to allocate instructions

basic_block_arena: SlotMap<BBId, BasicBlock>

An arena to allocate basic block info

first_block: Option<BBId>
Implementations

impl TacFunc

pub fn new(name: SmolStr, ty: Ty) -> TacFunc

pub fn new_untyped(name: SmolStr) -> TacFunc

pub fn starting_block(&self) -> Option<BBId>

impl TacFunc

Methods for manipulating instructions inside a function.

Note: All functions are infallible, and would panic when you use it with an ID that does not exist with it.

pub fn inst_new(&mut self, inst: Inst) -> InstId

Insert a new TAC into arena with no next instruction, without a proper basic block ID.

pub fn inst_exists(&self, inst: InstId) -> bool

pub fn inst_next(&self, inst: InstId) -> Option<InstId>

pub fn inst_prev(&self, inst: InstId) -> Option<InstId>

pub fn tac_get(&self, idx: InstId) -> &Tac

Gets the instruction with extra information around it as a Tac

pub fn tac_get_mut(&mut self, idx: InstId) -> &mut Tac

Gets a mutable reference of the instruction with extra information around it as a Tac

pub fn inst_get(&self, idx: InstId) -> &Inst

Get the instruction body

pub fn inst_get_mut(&mut self, idx: InstId) -> &mut Inst

Get a mutable reference of the instruction body

pub fn inst_set_after(&mut self, after: InstId, inst: InstId)

Position this instruction after the given instruction.

pub fn inst_set_before(&mut self, before: InstId, inst: InstId)

Position this instruction before the given instruction.

pub fn inst_append_in_bb(&mut self, inst: InstId, bb: BBId)

Append the given instruction as the last instruction in basic block

pub fn inst_prepend_in_bb(&mut self, inst: InstId, bb: BBId)

Prepend the given instruction as the first instruction in basic block

pub fn inst_detach(&mut self, idx: InstId)

Detaches this instruction from the instruction chain.

pub fn inst_remove(&mut self, idx: InstId) -> Inst

Remove the given instruction

pub fn inst_connect(&mut self, tail: InstId, head: InstId)

Connect TAC instruction head to the place after tail.

Errors if tail does not exist or tail.next is not None. This function does not check for availability of head.

Panics

This function requires head and tail to be different. Panics if head == tail.

pub(crate) fn inst_split_off_after(&mut self, pos: InstId) -> Option<InstId>

Splits off instruction chain after position pos. Returns the index of head instruction of the latter chain.

Errors if pos does not exist or there is no code after tail.

pub fn all_inst_unordered(&self) -> impl Iterator<Item = (InstId, BBId, &Inst)>

Returns an iterator of all instructions inside this function, unordered.

pub fn inst_of_bb_iter(&self, bb: BBId) -> impl Iterator<Item = (InstId, &Inst)>

pub fn inst_iter(
&self,
start: Option<InstId>,
end: Option<InstId>
) -> impl Iterator<Item = (InstId, &Inst)>

impl TacFunc

Methods for playing with basic blocks

pub fn bb_new(&mut self) -> BBId

Insert a new basic block into this function

pub fn bb_exists(&self, idx: BBId) -> bool

pub fn bb_get(&self, idx: BBId) -> &BasicBlock

pub fn bb_get_mut(&mut self, idx: BBId) -> &mut BasicBlock

pub fn bb_get2_mut(
&mut self,
i1: BBId,
i2: BBId
) -> (&mut BasicBlock, &mut BasicBlock)

pub fn bb_set_first(&mut self, bb: BBId) -> Option<BBId>

Set the given block as the first block in function. Returns the previous first block.

pub fn bb_set_before(&mut self, before: BBId, bb: BBId)

pub fn bb_set_after(&mut self, after: BBId, bb: BBId)

pub fn bb_detach(&mut self, bb: BBId)

pub fn all_bb_unordered(&self) -> impl Iterator<Item = (BBId, &BasicBlock)>

pub fn bb_split_after(&mut self, inst: InstId, transfer_branches: bool) -> BBId

Split all instruction after inst into a new basic block. Returns the ID of that basic block.

Creates a new basic block new_bb

Move every instruction after inst and inside its basic block to new_bb

Move every jump instruction to new_bb if transfer_branches is true

pub fn bb_connect(&mut self, front: BBId, back: BBId) -> Branch

Concatenate basic block back into front and detaches back from basic blocks. Returns branch instruction inside front.

Move every instruction inside back into front

Remove all jump instruction inside front

Move all instrcution inside back into front

Detach back

pub fn bb_iter(&self) -> impl Iterator<Item = (BBId, &BasicBlock)>