Struct azuki_tac::builder::FuncEditor

pub struct FuncEditor<'a> {
    pub func: &'a mut TacFunc,
    current_bb_id: BBId,
    current_idx: Option<InstId>,
}

An editor attached to the given function for linear editing purposes.

Fields
func: &'a mut TacFunccurrent_bb_id: BBId

The basic block we're currently working on. Must be a valid basic block inside this function.

current_idx: Option<InstId>

The instruction index we're currently working on. New instructions will be inserted before or after this instruction, depending on the function we use.

This value MUST refer to an instruction inside current_bb. If this value is None, current_bb MUST be empty.

Implementations

impl<'a> FuncEditor<'a>

pub fn new(func: &'a mut TacFunc) -> FuncEditor<'a>

Create a new function editor with ABSOLUTELY nothing initialized.

PLEASE DEFINITELY REMEMBER TO INITIALIZE BEFORE PUTTING ANYTHING INSIDE

pub fn set_type(&mut self, ty: Ty)

pub fn current_bb_id(&self) -> BBId

Returns the current basic block this builder is working on.

pub fn current_bb(&self) -> &BasicBlock

pub fn current_bb_mut(&mut self) -> &mut BasicBlock

pub fn current_idx(&self) -> Option<InstId>

Returns the current instruction this builder is working on. If current_bb is empty, returns None.

pub fn current_inst(&self) -> Option<&Inst>

pub fn current_tac(&self) -> Option<&Tac>

pub fn current_inst_mut(&mut self) -> Option<&mut Inst>

pub fn current_tac_mut(&mut self) -> Option<&mut Tac>

pub fn new_bb(&mut self) -> BBId

Add an empty basic block into the function.

pub fn set_current_bb(&mut self, bb_id: BBId) -> bool

Set current basic block to bb_id. Also sets current_idx to the end of this basic block.

Returns whether the position was unchanged.

pub fn set_current_bb_start(&mut self, bb_id: BBId) -> bool

Set current basic block to bb_id. Also sets current_idx to the start of this basic block.

Returns whether the position was unchanged.

pub fn set_position_at_instruction(
&mut self,
inst_idx: InstId
) -> TacResult<bool>

Sets current basic block and instruction position at the position of the given instruction.

Returns whether the position was unchanged.

pub fn insert_after_current_place(&mut self, inst: Inst) -> InstId

Insert the given instruction after the current place. Returns the index to the inserted instruction (and also the SSA value it's related to).

If the current basic block is empty, the instruction is inserted as the only instruction of the basic block.

pub fn insert_before_current_place(&mut self, inst: Inst) -> InstId

Insert the given instruction before the current place. Returns the index to the inserted instruction (and also the SSA value it's related to).

If the current basic block is empty, the instruction is inserted as the only instruction of the basic block.

pub fn insert_at_end_of(&mut self, inst: Inst, bb_id: BBId) -> TacResult<InstId>

Insert the given instruction at the end of the given basic block.

pub fn insert_at_start_of(
&mut self,
inst: Inst,
bb_id: BBId
) -> TacResult<InstId>

Insert the given instruction at the start of the given basic block.

pub fn put_inst_after_current_place(&mut self, idx: InstId)

Attach the free-standing instruction to the place after [current_idx], and advance one instruction forward.

Panics

Panics when the instruction is not free-standing (inst.prev or inst.next is not None).

fn put_inst_before_current_place(&mut self, idx: InstId)

Attach the free-standing instruction to the place before [current_idx], and advance one instruction back.

Panics

Panics when the instruction is not free-standing (inst.prev or inst.next is not None).

pub fn insert_phi(&mut self, bb_id: BBId, ty: Ty) -> Result<InstId, Error>

pub fn move_forward(&mut self) -> bool

Move one instruction forward. Returns whether the move was successful. If this function returns true, [current_idx] and functions related to it are guaranteed to return Some as long as .

pub fn move_backward(&mut self) -> bool

Move one instruction backward. Returns whether the move was successful. If this function returns true, [current_idx] and functions related to it are guaranteed to return Some.

pub fn remove_current(&mut self) -> Option<Inst>